Gift AI
This is a code for a gift AI. This code is python. I made most of this code on Replit. I also used chat GPT for the code. Those are my credits. It is randomized so you won't have to choose your friend, family, and co-workers gift for the holidays.